Commit running version of code that works bug fix to 32 bit integer overflow
Former-commit-id: de12368e7d1151cac20b84a711028f11ae2a9243
Former-commit-id: 4d8bdcd8f13eb83ac4f1efeced4b2aa3562cb2b4 [formerly df27c2ac993e5de01d336d2a66057c5bb3e0c46a] [formerly 98456aa824a3806548173f3a02e9d7f27a03e52e [formerly 566f2d3c073d5c62988a36b63e0e418ef41d2ba9]]
Former-commit-id: 786c0a5f820c8e971603a12b7a246a5f1c4ab2a8 [formerly 43acf3d2dc866c6d17f2da0e8f44e3c21b1ffa0e]
Former-commit-id: c70364bf0fa6500d2ac27b382cba32fc152b261c

+++ src/applications/model/packet-sink.cc
@@ -106,7 +106,7 @@ PacketSink::~PacketSink()
   NS_LOG_FUNCTION (this);
 }
 
-uint32_t PacketSink::GetTotalRx () 
+uint64_t PacketSink::GetTotalRx () 
 {
   NS_LOG_FUNCTION (this);
